+++ src/Command/Command.php
@@ -2,11 +2,13 @@
 
 namespace Consolly\Command;
 
+use Consolly\Option\OptionInterface;
+
 /**
  * Class Command represents helpful implementation of the {@link CommandInterface}.
  *
- * WARNING: You must define values for every variable of this class because it have no default value.
- * Otherwise when trying to access the variable, an exception will be thrown because the variable is not initialized.
+ * WARNING: You must define values for every variable of this class because it has no default value.
+ * Otherwise, when trying to access the variable, an exception will be thrown because the variable is not initialized.
  * For example, you can define values for the variables in the constructor.
  *
  * @package Consolly\Command
@@ -36,11 +38,41 @@ class Command implements CommandInterface
     }
 
     /**
+     * Sets command name.
+     *
+     * @param string $name
+     */
+    public function setName(string $name): void
+    {
+        $this->name = $name;
+    }
+
+    /**
      * @inheritdoc
      */
     public function getOptions(): array
     {
         return $this->options;
+    }
+
+    /**
+     * Sets commands options
+     *
+     * @param array $options
+     */
+    public function setOptions(array $options): void
+    {
+        $this->options = $options;
+    }
+
+    /**
+     * Adds the option to the options array.
+     *
+     * @param OptionInterface $option
+     */
+    public function addOption(OptionInterface $option): void
+    {
+        $this->options[] = $option;
     }
 
     /**