+++ docs/changelog.md
@@ -1,1 +1,18 @@
-Will be added later.
+# Changelog
+Here is a list of versions with changes from first to last.
+
+## v2.0.0
+- Added more abstraction layers.
+  - Added a DistributorInterface with implementation.
+  - Added a SourceInterface with implementation.
+- Added more expansion options.
+  - Now, command is not abstract class. Now you should use CommandInterface or Command class.
+  - Now, option is not abstract class. Now you should use OptionInterface or Option class.
+- Added tests.
+- Added documentation.
+
+## v1.0.0
+First release.
+
+- Added main class with all logic.
+- Added command abstract class, and option abstract class.